Migration OWSI-Core : Wicket 6.1.1, Wicket 6.0.0 (premier jet... ça va péter dans tous les sens)
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@1607 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core-components/owsi-core-component-wicket-more/src/main/java/fr/openwide/core/wicket/more/markup/html/navigation/paging/AjaxBootstrapPagingNavigation.java
@@ -13,7 +13,7 @@ public class AjaxBootstrapPagingNavigation extends BootstrapPagingNavigation {
 	}
 	
 	@Override
-	protected Link<?> newPagingNavigationLink(String id, IPageable pageable, int pageIndex) {
+	protected Link<?> newPagingNavigationLink(String id, IPageable pageable, long pageIndex) {
 		return new AjaxBootstrapPagingNavigationLink(id, pageable, pageIndex).setAutoEnable(false);
 	}
